complementando com remover e editar

+++ src/main/java/com/market/negocio/IEmpresa.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.market.model.Empresa;
 
-public interface IEmpresa extends JpaRepository<Empresa, Long>{
+public interface IEmpresa extends JpaRepository<Empresa, Integer>{
 
 }
